+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Support for dnspython 2.0, Salt 3002, Python 3.9
+- Allow to customize file path for private key and certificate files
 
 ## [1.1.0] - 2020-08-04
 ### Added